requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-tools-duckduckgo